requirements.txt
selenium==4.29.0
python-dotenv==0.21.0
